code/.ipynb_checkpoints/requirements-checkpoint.txt
numpy
numba
pytorch-lightning
Cython
asteroid
sagemaker-inference
PyYAML
torchaudio